# swagger.yaml
swagger: "2.0"
info:
  title: NENS API documentation
  description: Contains all available API endpoints in this codebase
  version: "1.0.0"
  termsOfService: "http://swagger.io/terms/"
host: localhost:3000
basePath: /
schemes:
  - https
paths:
...
  /articles/{id}:

    # GET /articles/{id} => Read Article
    get:
      tags:
        - Article Module
      summary: Read Article
      description: API to Read an Article
      parameters:
      - name: id
        type: number
        in: path
        description: Id of the Article
        required: true
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/ArticleRes"
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

    # PUT /articles/{id} => Update Article
    put:
      security:
        - Bearer: []
      tags:
        - Article Module
      summary: Update Article
      description: API to Update an Article
      consumes:
        - multipart/form-data
      parameters:
        - name: id
          type: number
          in: path
          description: Id of the Article
          required: true
        - in: body
          name: body
          description: updateArticle() Payload
          required: true
          schema:
            $ref: "#/definitions/ArticleReq"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/BasicRes"
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

    # DELETE /articles/{id} => Delete Article
    delete:
      security:
        - Bearer: []
      tags:
        - Article Module
      summary: Delete Article
      description: API to Delete an Article
      parameters:
        - name: id
          type: number
          in: path
          description: Id of the Article
          required: true
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: "#/definitions/BasicRes"
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

  # ! AUTH
  # GET /auth/{id} => Read Avatar
...
  /projects/{id}:

    # PUT /projects/{id} => Update Project
    put:
      security:
        - Bearer: []
      tags:
        - Project Module
      summary: Update Project
      description: API to Update a Project
      consumes:
        - multipart/form-data
      parameters:
        - name: id
          type: number
          in: path
          description: Id of the Project
          required: true
        - in: body
          name: body
          description: updateProject() Payload
          required: true
          schema:
            $ref: "#/definitions/ProjectReq"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/BasicRes"
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

    # DELETE /projects/{id} => Delete Project
    delete:
      security:
        - Bearer: []
      tags:
        - Project Module
      summary: Delete Project
      description: API to Delete a Project
      parameters:
        - name: id
          type: number
          in: path
          description: Id of the Project
          required: true
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: "#/definitions/BasicRes"
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error

  # ! USERS
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  ArticleReq:
    properties:
      name:
        type: string
        description: article name
      text:
        type: string
        description: article content
      image:
        type: string
        description: article image
      alt:
        type: string
        description: alternative text of article image
      url:
        type: string
        description: article url
      cat:
        type: string
        description: article category

  # AUTH REQUEST
...
  ProjectReq:
    properties:
      name:
        type: string
        description: project name
      description:
        type: string
        description: project content
      image:
        type: string
        description: project image
      alt:
        type: string
        description: alternative text of project image
      url:
        type: string
        description: project url
      cat:
        type: string
        description: project category

  # USER REQUEST
...
  ArticleRes:
    type: object
    properties:
      id:
        type: number
        description: article id
      name:
        type: string
        description: article name
      text:
        type: string
        description: article content
      image:
        type: string
        description: article image
      alt:
        type: string
        description: alternative text of article image
      url:
        type: string
        description: article url
      likes:
        type: string
        description: article likes
      cat:
        type: string
        description: article category
      createdAt:
        type: string
        format: date
        description: article creation
      upatedAt:
        type: string
        format: date
        description: article update

  # AUTH RESPONSE
...
  BasicRes:
    properties:
      status:
        type: integer
        description: status code
      message:
        type: string
        description: confirm message

  # GALLERY RESPONSE
...